fix: Allow matching remote URL for .github repo (fix regex)
The regex was disallowing repos with dots in their name. This commit adds
`.` as an allowable character, but includes some other changes as well:

- Remove commas from regex (not valid syntax)
- Replace `{0,1}` with `?` and remove some unneeded groupings
- Remove owner capture, since it wasn't being used
- Allow hostname and path prefix to just be matched by an undifferentiated
  `.*` since we don't really care about the structure there.
- Both anchor the regex and switch to fullmatch, which is redundant but
  both are best-practice unless the situation calls for something more
  relaxed.
- Rename regex var to better represent its purpose.

Arguably we don't really need the git/https scheme matching, but there's a
small risk that this could be pointed at a repo cloned from a directory,
and that directory might have a generic name like "workspace" or
"checkout". Until we have a need to relax this constraint, we should
probably keep it in place.